connected users' favorites to the database

+++ frontend/src/app/components/pages/favorites-page/favorites-page.component.ts
@@ -1,22 +1,43 @@
-import { Component } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FavoritesService } from '../../../services/favorites.service';
 import { Favorites } from '../../../shared/models/Favorites';
 import { FavoritePet } from '../../../shared/models/FavoritePet';
+import { Subscription } from 'rxjs';
+
 
 @Component({
-  selector: 'app-favorites-page',
-  templateUrl: './favorites-page.component.html',
-  styleUrl: './favorites-page.component.css'
+    selector: 'app-favorites-page',
+    templateUrl: './favorites-page.component.html',
+    styleUrls: ['./favorites-page.component.css']
 })
-export class FavoritesPageComponent {
-  favorites!: Favorites;
-  constructor(private favoritesService:FavoritesService) {
-    this.favoritesService.getFavoritesObservable().subscribe((favorites) => {
-      this.favorites = favorites;
-    })
-  }
+export class FavoritesPageComponent implements OnInit, OnDestroy {
+    favorites!: Favorites;
+    private favoritesSubscription: Subscription;
 
-  removeFromFavorites(favoritePet:FavoritePet){
-    this.favoritesService.removeFromFavorites(favoritePet.pet.id);
-  }
-}
+    constructor(private favoritesService: FavoritesService) {
+      this.favoritesSubscription = this.favoritesService.getFavoritesObservable().subscribe((favorites) => {
+        this.favorites = favorites;
+      });
+     };
+
+    ngOnInit(): void {
+        this.favoritesService.loadFavorites();
+        this.favoritesSubscription = this.favoritesService.getFavoritesObservable().subscribe((favorites) => {
+            this.favorites = favorites;
+        });
+    };
+
+    ngOnDestroy(): void {
+        if (this.favoritesSubscription) {
+            this.favoritesSubscription.unsubscribe();
+        };
+    };
+
+    removeFromFavorites(favoritePet: FavoritePet) {
+        if (favoritePet && favoritePet.pet && favoritePet.pet.id) {
+            this.favoritesService.removeFromFavorites(favoritePet.pet.id);
+        } else {
+            console.error('Invalid favoritePet object or pet id');
+        };
+    };
+};
